+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix inconsistency between JIT and VM evaluator when performing interval
   evaluation of `not([NAN, NAN])`.
 - Propagate `NAN` values through `and` and `or` operations on intervals.
+- Remove `Grad::to_rgb` in favor of handling it at the image level
 
 # 0.3.4
 - Add `GenericVmFunction::simplify_with` to simultaneously simplify a function